Add functions for simulating single-trial encoding and rsa models
Description:

`params.R` contains a funciton, `new_params()`, which creates a parameters list. Parameter lists contain all parameters needed to specify a simulation, and are passed to functions that carry out the simulation.

`utils.R` contains functions that are called by `new_params()` or in the process of simulation.

Some example code snippets for a single iteration of a simulation are provided in example.R